refactor: move projects injection token to data
