+++ src/app/resume-page/data/adapt-json-resume-work.ts
@@ -1,7 +1,7 @@
 import { inject, InjectionToken } from '@angular/core'
 import { DateRange } from './date-range'
 import { RELATIVIZE_PRODUCTION_URL } from '@/common/relativize-production-url'
-import { JSON_RESUME_PROJECTS } from '../projects-section/json-resume-projects'
+import { JSON_RESUME_PROJECTS } from './json-resume-projects'
 import { ADAPT_JSON_RESUME_PROJECT } from './adapt-json-resume-project'
 import { JsonResumeWorkItem } from '../json-resume/types'
 import { Experience } from './experience'
